refactor: move space frames to separate module
