+++ src/coordinax/_coordinax_space_frames/__init__.py
@@ -1,0 +1,14 @@
+"""Frames for Astronomy."""
+
+__all__ = ["ICRS", "Galactocentric"]
+
+from .frame_transforms import *
+from .galactocentric import Galactocentric
+from .icrs import ICRS
+
+# Interoperability. Importing this module will register interop frameworks.
+# isort: split
+from . import _interop
+
+# clean up namespace
+del _interop
