+++ src/coordinax/_coordinax_space_frames/_interop/__init__.py
@@ -1,0 +1,8 @@
+"""Interoperability."""
+
+__all__: list[str] = []
+
+from .optional_deps import OptDeps
+
+if OptDeps.ASTROPY.installed:
+    from . import interop_astropy  # noqa: F401
