+++ src/coordinax/_coordinax_space_frames/_interop/interop_astropy/__init__.py
@@ -2,4 +2,4 @@
 
 __all__: list[str] = []
 
-from . import constructors, converters, distance  # noqa: F401
+from . import frames  # noqa: F401
