+++ src/coordinax/_coordinax_space_frames/_interop/optional_deps.py
@@ -1,0 +1,11 @@
+"""Optional dependencies. Internal use only."""
+
+__all__ = ["OptDeps"]
+
+from optional_dependencies import OptionalDependencyEnum, auto
+
+
+class OptDeps(OptionalDependencyEnum):  # type: ignore[misc]  # pylint: disable=invalid-enum-extension
+    """Optional dependencies for ``coordinax_space_frames``."""
+
+    ASTROPY = auto()
