+++ src/coordinax/frames.py
@@ -1,13 +1,17 @@
 """Reference frames and transformations between them."""
 
+from . import _coordinax_space_frames
+from ._src.frames.api import frame_transform_op
+from ._src.frames.base import AbstractReferenceFrame
+
+# isort: split
+from ._coordinax_space_frames import *  # noqa: F403
+
 __all__ = [
     "frame_transform_op",
-    # Frames
     "AbstractReferenceFrame",
-    "ICRS",
-    "Galactocentric",
 ]
+__all__ += _coordinax_space_frames.__all__
 
-from ._src.frames.api import frame_transform_op
-from ._src.frames.astro_frames import ICRS, Galactocentric
-from ._src.frames.base import AbstractReferenceFrame
+# clean up namespace
+del _coordinax_space_frames
